Removed GZip and instead use the default one from .Net. Refactor for saving and loading GRFs. Cleaned up all GRF save methods and now return a result of the save instead of throwing exceptions. Better error report when extracting files from a GRF. Better progress updates when saving a GRF. Reintroduced file association with a cleaner interface. Simplified methods related to setting encryption key. Added the lub effect textures for the map resource extractions tree. Fixed GRF reloading mixed behaviors when saving a GRF (this now only applies for Thor and RGZ formats). Fixed an issue with non-closed streams when repacking a GRF and canceling. Saving a GRF will now create a temporary file within the current folder rather than using the %AppData% path. This prevents copying the GRF from one drive to another (only applies for FileCopy mode). Renamed "GrfSave" to "FileCopy"; renamed "QuickMerge" to "FileEdit". Encrypting GRF entries can now use the FileEdit mode rather than always doing a file copy. This was especially noticeable for large GRFs.

+++ GRF/Threading/GrfThreadExtract.cs
@@ -7,6 +7,23 @@ using GRF.Core;
 using Utilities.Extension;
 
 namespace GRF.Threading {
+	public enum ExtractionFailReason {
+		CannotWriteFileToDisk,
+		CorruptedOrEncrypted,
+		GravityEncrypted,
+		SourceFileMissing,
+	}
+
+	public class ExtractionResult {
+		public FileEntry Entry;
+		public ExtractionFailReason Reason;
+
+		public ExtractionResult(FileEntry entry, ExtractionFailReason reason) {
+			Entry = entry;
+			Reason = reason;
+		}
+	}
+
 	/// <summary>
 	/// This class extracts and copy the files from the GRF from a given range
 	/// It's used to optimize the data transfer.
@@ -14,6 +31,7 @@ namespace GRF.Threading {
 	public class GrfThreadExtract : GrfWriterThread<FileEntry> {
 		private const int _bufferSize = 8388608;
 		private readonly StreamReadBlockInfo _srb = new StreamReadBlockInfo(_bufferSize);
+		public List<ExtractionResult> ErrorResults = new List<ExtractionResult>();
 
 		public override void Start() {
 			new Thread(_start) { Name = "GRF - Extract files thread " + StartIndex }.Start();
@@ -40,7 +58,7 @@ namespace GRF.Threading {
 
 					while (toIndex < indexMax) {
 						fromIndex = toIndex;
-						data = _srb.ReadMisaligned(sortedEntries, out toIndex, fromIndex, indexMax, originalStream.Value);
+						data = _srb.ReadMisaligned(sortedEntries, out toIndex, fromIndex, indexMax, originalStream.Value, out _);
 
 						for (int i = fromIndex; i < toIndex; i++) {
 							if (_grfData.IsCancelling)
@@ -53,7 +71,7 @@ namespace GRF.Threading {
 							entry.GrfEditorDecryptRequested(data);
 							entry.DesDecrypt(data, (int) entry.TemporaryOffset);
 							entry.Align(data, (int) entry.TemporaryOffset, out dataTmp);
-
+							
 							if (entry.RelativePath.GetExtension() == null) {
 								// Fix: 2018-09-26
 								// Sometimes Gravity sent the directory link as a file inside the GRF, ignore those when extraccting.
@@ -64,8 +82,11 @@ namespace GRF.Threading {
 							}
 
 							try {
-								if (entry.Flags.HasFlag(EntryType.GravityEncryptedFile))
-									dataTmp = null;
+								if (entry.Flags.HasFlag(EntryType.GravityEncryptedFile)) {
+									Error = true;
+									ErrorResults.Add(new ExtractionResult(entry, ExtractionFailReason.GravityEncrypted));
+									continue;
+								}
 								else if ((entry.Flags & EntryType.LZSS) == EntryType.LZSS)
 									dataTmp = Compression.LzssDecompress(dataTmp, entry.SizeDecompressed);
 								else if ((entry.Flags & EntryType.RawDataFile) == EntryType.RawDataFile)
@@ -76,37 +97,55 @@ namespace GRF.Threading {
 									dataTmp = Compression.Decompress(dataTmp, entry.SizeDecompressed);
 
 								if (dataTmp != null) {
-									using (FileStream fs = new FileStream(entry.ExtractionFilePath, FileMode.Create, FileAccess.Write, FileShare.Read))
+									using (FileStream fs = new FileStream(entry.ExtractionFilePath, FileMode.Create, FileAccess.Write, FileShare.None, 4096, FileOptions.WriteThrough))
 										fs.Write(dataTmp, 0, dataTmp.Length);
 								}
 							}
+							catch (IOException err) {
+								Error = true;
+								ErrorResults.Add(new ExtractionResult(entry, ExtractionFailReason.CannotWriteFileToDisk));
+								Exception = err;
+							}
 							catch (Exception err) {
 								Error = true;
-								Exception = new Exception("#File: " + entry.RelativePath, err);
+								ErrorResults.Add(new ExtractionResult(entry, ExtractionFailReason.CorruptedOrEncrypted));
+								Exception = err;
 							}
 							finally {
 								entry.TemporaryOffset = 0;
+								NumberOfFilesProcessed++;
 							}
-
-							NumberOfFilesProcessed++;
 						}
 					}
 
 					foreach (FileEntry entryCopy in sortedEntriesAdded) {
-						if (entryCopy.SourceFilePath != entryCopy.ExtractionFilePath) {
-							if (_grfData.IsCancelling)
-								return;
+						try {
+							if (entryCopy.SourceFilePath != entryCopy.ExtractionFilePath) {
+								if (_grfData.IsCancelling)
+									return;
 
-							if (IsPaused)
-								Pause();
+								if (IsPaused)
+									Pause();
 
-							if (File.Exists(entryCopy.ExtractionFilePath))
-								File.Delete(entryCopy.ExtractionFilePath);
+								if (File.Exists(entryCopy.ExtractionFilePath))
+									File.Delete(entryCopy.ExtractionFilePath);
 
-							File.Copy(entryCopy.SourceFilePath, entryCopy.ExtractionFilePath);
+								if (!File.Exists(entryCopy.SourceFilePath)) {
+									ErrorResults.Add(new ExtractionResult(entryCopy, ExtractionFailReason.SourceFileMissing));
+									continue;
+								}
+
+								File.Copy(entryCopy.SourceFilePath, entryCopy.ExtractionFilePath);
+							}
 						}
-
-						NumberOfFilesProcessed++;
+						catch (Exception err) {
+							Error = true;
+							ErrorResults.Add(new ExtractionResult(entryCopy, ExtractionFailReason.CannotWriteFileToDisk));
+							Exception = err;
+						}
+						finally {
+							NumberOfFilesProcessed++;
+						}
 					}
 				}
 			}
